# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mentora AI API",
    "description": "API versionada do Mentora AI.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/classes": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "List User Classrooms",
        "description": "List classrooms the authenticated user is enrolled in.",
        "operationId": "list_user_classrooms_api_v1_classes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClassroomResponse"
                  },
                  "type": "array",
                  "title": "Response List User Classrooms Api V1 Classes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Create Classroom",
        "description": "Create a new classroom and enroll the creator as educator.",
        "operationId": "create_classroom_api_v1_classes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/join": {
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Join Classroom By Code",
        "description": "Join a classroom using the 6-character invitation code.",
        "operationId": "join_classroom_by_code_api_v1_classes_join_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinClassroomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{classroom_id}": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "Get Classroom Detail",
        "description": "Get classroom details and verify enrollment.",
        "operationId": "get_classroom_detail_api_v1_classes__classroom_id__get",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "classes"
        ],
        "summary": "Update Classroom",
        "description": "Update classroom settings, including pedagogical mode (educator only).",
        "operationId": "update_classroom_api_v1_classes__classroom_id__patch",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassroomRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{classroom_id}/members": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "List Classroom Members",
        "description": "List enrolled members (educator only).",
        "operationId": "list_classroom_members_api_v1_classes__classroom_id__members_get",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassroomMemberResponse"
                  },
                  "title": "Response List Classroom Members Api V1 Classes  Classroom Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{classroom_id}/radar": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "Get Learning Gap Radar",
        "description": "Get aggregated difficulty radar and fallback alerts without student PII (educator only).",
        "operationId": "get_learning_gap_radar_api_v1_classes__classroom_id__radar_get",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningGapRadarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{classroom_id}/flashcards": {
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Create Flashcard",
        "description": "Create a personal study flashcard for the current student.",
        "operationId": "create_flashcard_api_v1_classes__classroom_id__flashcards_post",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlashcardRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashcardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "List Student Flashcards",
        "description": "List the current student's flashcards in this classroom.",
        "operationId": "list_student_flashcards_api_v1_classes__classroom_id__flashcards_get",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlashcardResponse"
                  },
                  "title": "Response List Student Flashcards Api V1 Classes  Classroom Id  Flashcards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{classroom_id}/flashcards/{flashcard_id}": {
      "patch": {
        "tags": [
          "classes"
        ],
        "summary": "Update Flashcard Review Status",
        "description": "Update a flashcard's review status (pending vs mastered).",
        "operationId": "update_flashcard_review_status_api_v1_classes__classroom_id__flashcards__flashcard_id__patch",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Classroom Id"
            }
          },
          {
            "name": "flashcard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flashcard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlashcardStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashcardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClassroomResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "created_by_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Created By User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "pedagogical_mode": {
            "$ref": "#/components/schemas/PedagogicalMode"
          },
          "knowledge_base_id": {
            "type": "string",
            "format": "uuid",
            "title": "Knowledge Base Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassroomRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "created_by_user_id",
          "name",
          "code",
          "description",
          "pedagogical_mode",
          "knowledge_base_id",
          "created_at"
        ],
        "title": "ClassroomResponse"
      },
...
      "CreateClassroomRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "pedagogical_mode": {
            "$ref": "#/components/schemas/PedagogicalMode",
            "default": "EXPLANATION"
          },
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Base Id",
            "description": "ID da Base de Conhecimento vinculada; se omitido, usa a padrão do tenant."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateClassroomRequest"
      },
      "CreateFlashcardRequest": {
        "properties": {
          "front_prompt": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Front Prompt"
          },
          "back_answer": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Back Answer"
          },
          "source_message_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Message Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "front_prompt",
          "back_answer"
        ],
        "title": "CreateFlashcardRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JoinClassroomRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 10,
            "minLength": 4,
            "title": "Code",
            "description": "Código curto de convite da turma (ex: 'FIS104' ou 'K9M2W7')"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "JoinClassroomRequest"
      },
...
      "UpdateClassroomRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "pedagogical_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PedagogicalMode"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateClassroomRequest"
      },
...
    }
  }
}